move function to test to its own module
